requirements.txt
aiofile==3.8.8
aiofiles==23.2.1
aiogram==3.6.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.2.0
caio==0.9.17
certifi==2024.2.2
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.1
executing==2.0.1
fastapi-cli==0.0.5
filelock==3.14.0
frozenlist==1.4.1
fsspec==2024.5.0
gigachat==0.1.27
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
ipython==8.24.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
lightning-utilities==0.11.2
llvmlite==0.42.0
magic-filter==1.0.12
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
openai==1.30.2
openai-whisper==20231117
orjson==3.10.7
packaging==24.0
paramiko==3.4.1
parso==0.8.4
peewee==3.17.5
pexpect==4.9.0
pillow==10.4.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.7.1
pydantic-extra-types==2.9.0
pydantic-settings==2.4.0
pydantic_core==2.18.2
pydub==0.25.1
Pygments==2.18.0
PyNaCl==1.5.0
PyPDF2==3.0.1
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-lightning==2.2.5
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.2
rich==13.7.1
semantic-version==2.10.0
setuptools==72.1.0
setuptools-rust==1.9.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.37.2
svgwrite==1.4.3
sympy==1.12
tiktoken==0.7.0
tomli==2.0.1
torch==2.3.0
torchmetrics==1.4.0.post0
tqdm==4.66.4
traitlets==5.14.3
Tree==0.2.4
triton==2.3.1
typer==0.12.3
typing_extensions==4.11.0
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.5
uvloop==0.19.0
watchfiles==0.23.0
wcwidth==0.2.13
websockets==12.0
yarl==1.9.4